Moved Divine Image tests
